+++ image/nodejs.sh
@@ -4,4 +4,7 @@ source /pd_build/buildconfig
 set -x
 
 ## Install Node.js (also needed for Rails asset compilation)
-apt-get install -y nodejs
+minimal_apt_get_install -y nodejs npm
+if [[ ! -e /usr/bin/node ]]; then
+	ln -s /usr/bin/nodejs /usr/bin/node
+fi