+++ image/ruby2.2.sh
@@ -18,5 +18,5 @@ update-alternatives \
 	--slave /usr/share/man/man1/irb.1.gz irb.1.gz /usr/share/man/man1/irb2.2.*.gz \
 	--slave /usr/share/man/man1/rake.1.gz rake.1.gz /usr/share/man/man1/rake2.2.*.gz \
 	--slave /usr/share/man/man1/ri.1.gz ri.1.gz /usr/share/man/man1/ri2.2.*.gz
-gem2.2 install rake bundler --no-rdoc --no-ri
+gem2.2 install rake bundler --no-document
 /pd_build/ruby-finalize.sh